modules/WaddleBot-Listener/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
pydal==20240428.2
python-dotenv==1.0.1
redis==5.0.4
requests==2.31.0
urllib3==2.2.1
